+++ src/Metadata/GoodReads/GoodReadsMatch.php
@@ -49,7 +49,7 @@ class GoodReadsMatch extends BaseMatch
             return $this->cache->loadCache($cacheFile);
         }
         $url = str_replace('{query}', rawurlencode($query), self::QUERY_URL);
-        $result = file_get_contents($url, false, $this->context);
+        $result = $this->getUrlContent($url);
         $matched = $this->parseSearchPage($query, $result);
         $this->cache->saveCache($cacheFile, $matched);
         usleep(self::SLEEP_TIME);
@@ -180,7 +180,7 @@ class GoodReadsMatch extends BaseMatch
         }
         // https://www.goodreads.com/author/list/123.Author_Name?per_page=100
         $url = self::AUTHOR_URL . $authorId . '?per_page=100';
-        $result = file_get_contents($url, false, $this->context);
+        $result = $this->getUrlContent($url);
         $parsed = $this->parseAuthorPage($authorId, $result);
         // @todo remove other authors here?
         $entity = $parsed;
@@ -255,7 +255,7 @@ class GoodReadsMatch extends BaseMatch
         }
         // https://www.goodreads.com/series/123.Series_Name
         $url = self::SERIES_URL . $seriesId;
-        $result = file_get_contents($url, false, $this->context);
+        $result = $this->getUrlContent($url);
         $parsed = $this->parseSeriesPage($seriesId, $result);
         $this->cache->saveCache($cacheFile, $parsed);
         usleep(self::SLEEP_TIME);
@@ -358,7 +358,7 @@ class GoodReadsMatch extends BaseMatch
         }
         // https://www.goodreads.com/book/show/123.Book_Title
         $url = static::link($bookId);
-        $result = file_get_contents($url, false, $this->context);
+        $result = $this->getUrlContent($url);
         $result = $this->parseBookPage($bookId, $result);
         $entity = json_decode($result, true, 512, JSON_THROW_ON_ERROR);
         $this->cache->saveCache($cacheFile, $entity);