+++ src/Metadata/OpenLibrary/OpenLibraryMatch.php
@@ -49,7 +49,7 @@ class OpenLibraryMatch extends BaseMatch
         }
         // https://openlibrary.org/dev/docs/api/authors
         $url = 'https://openlibrary.org/search/authors.json?q=' . rawurlencode($query);
-        $results = file_get_contents($url, false, $this->context);
+        $results = $this->getUrlContent($url);
         $matched = json_decode($results, true);
         $this->cache->saveCache($cacheFile, $matched);
         usleep(parent::SLEEP_TIME);
@@ -105,7 +105,7 @@ class OpenLibraryMatch extends BaseMatch
         //$url = 'https://openlibrary.org/authors/' . $matchId . '/works.json?limit=' . $limit;
         // generic search returns 'docs' but author search returns 'entries'
         $url = 'https://openlibrary.org/search.json?author=' . $matchId . '&fields=key,type,title,edition_count,first_publish_year,number_of_pages_median,author_name,author_key';
-        $results = file_get_contents($url, false, $this->context);
+        $results = $this->getUrlContent($url);
         $matched = json_decode($results, true);
         $this->cache->saveCache($cacheFile, $matched);
         usleep(parent::SLEEP_TIME);
@@ -129,11 +129,11 @@ class OpenLibraryMatch extends BaseMatch
         }
         // https://openlibrary.org/dev/docs/api/search
         $url = 'https://openlibrary.org/search.json?title=' . rawurlencode($query) . '&author=' . rawurlencode((string) $authorName) . '&fields=key,type,title,edition_count,first_publish_year,number_of_pages_median,author_name,author_key';
-        $results = file_get_contents($url, false, $this->context);
+        $results = $this->getUrlContent($url);
         $matched = json_decode($results, true);
         if (empty($matched) || empty($matched['docs'])) {
             $url = 'https://openlibrary.org/search.json?title=' . rawurlencode($query) . '&fields=key,type,title,edition_count,first_publish_year,number_of_pages_median,author_name,author_key';
-            $results = file_get_contents($url, false, $this->context);
+            $results = $this->getUrlContent($url);
             $matched = json_decode($results, true);
         }
         $this->cache->saveCache($cacheFile, $matched);
@@ -156,7 +156,7 @@ class OpenLibraryMatch extends BaseMatch
         }
         // https://openlibrary.org/dev/docs/api/authors
         $url = self::AUTHOR_URL . $authorId . '.json';
-        $result = file_get_contents($url, false, $this->context);
+        $result = $this->getUrlContent($url);
         $entity = json_decode($result, true);
         $this->cache->saveCache($cacheFile, $entity);
         usleep(parent::SLEEP_TIME);
@@ -181,7 +181,7 @@ class OpenLibraryMatch extends BaseMatch
         }
         // https://openlibrary.org/dev/docs/api/books
         $url = static::link($workId) . '.json';
-        $result = file_get_contents($url, false, $this->context);
+        $result = $this->getUrlContent($url);
         $entity = json_decode($result, true);
         $this->cache->saveCache($cacheFile, $entity);
         usleep(parent::SLEEP_TIME);